+++ src/main/kotlin/com/atlassian/performance/tools/report/api/result/InteractionStats.kt
@@ -2,7 +2,7 @@ package com.atlassian.performance.tools.report.api.result
 
 import java.time.Duration
 
-data class InteractionStats(
+class InteractionStats(
     internal val cohort: String,
     internal val sampleSizes: Map<String, Long>?,
     internal val centers: Map<String, Duration>?,